Change ifconfig -> ip addr show
The ip command has superseded ifconfig in recent versions of Linux, which
means ifconfig is not likely to be found installed on newer machines.

Also en0 is not necessarily the name of the default interface, mine was
annoyingly named eno1.

To hopefully help with this I've changed ifconfig -> ip addr and removed
the interface name. We already filter out localhost with the following
regex, so this ends up with the first non-localhost ip address found by
ip addr.